// apps/mobile/src/components/ui/toast/types.ts
export interface ToastColors {
	"background-primary": string;
	"background-secondary": string;
	"text-primary": string;
	"text-secondary": string;
	"text-tertiary": string;
	"border-secondary": string;
	success: string;
	error: string;
	warning: string;
	info: string;
	rich: {
		success: {
			background: string;
			foreground: string;
			border: string;
		};
		error: {
			background: string;
			foreground: string;
			border: string;
		};
		warning: {
			background: string;
			foreground: string;
			border: string;
		};
		info: {
			background: string;
			foreground: string;
			border: string;
		};
	};
}

export interface ToastColorsConfig {
	light: ToastColors;
	dark: ToastColors;
}